Kill model only if needed

+++ pollinator/process_msg.py
@@ -34,6 +34,9 @@ class BackgroundCommand:
         logging.info("killed")
 
 
+loaded_model = None
+
+
 class RunningCogModel:
     def __init__(self, image, output_path):
         self.image = image
@@ -47,12 +50,21 @@ class RunningCogModel:
         logging.info(f"Initializing cog command: {self.cog_cmd}")
 
     def __enter__(self):
-        logging.info("Starting cog model")
+        global loaded_model
+        if loaded_model == self.image:
+            logging.info(f"Model already loaded: {self.image}")
+            return
+        if loaded_model is not None:
+            logging.info(f"Killing previous model ({loaded_model})")
+            os.system("docker kill cogmodel")
+
+        logging.info(f"Starting {self.image}: {self.cog_cmd}")
         os.system(self.cog_cmd)
+        loaded_model = self.image
 
     def __exit__(self, type, value, traceback):
-        logging.info(f"Killing {self.image}")
-        os.system("docker kill cogmodel")
+        # we leave the model running in case the next request needs the same model
+        pass
 
 
 def process_message(message):
@@ -88,9 +100,8 @@ def process_message(message):
 
 def prepare_output_folder(output_path):
     logging.info(f"Mounting output folder: {output_path}")
-    shutil.rmtree(output_path, ignore_errors=True)
     os.makedirs(output_path, exist_ok=True)
-
+    clean_folder(output_path)
     write_folder(output_path, "done", "false")
     write_folder(output_path, "time_start", str(int(time.time())))
 
@@ -133,6 +144,18 @@ def write_folder(path, key, value):
     os.makedirs(path, exist_ok=True)
     with open(f"{path}/{key}", "w") as f:
         f.write(value)
+
+
+def clean_folder(folder):
+    for filename in os.listdir(folder):
+        file_path = os.path.join(folder, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as e:
+            print("Failed to delete %s. Reason: %s" % (file_path, e))
 
 
 # if __name__ == "__main__":
